optimize deletion with indexed tombstones
For hashes over a threshold size, deletion will trigger indexing of
the position of ordered keys.  On deletion, keys will be tombstoned.
When the number of tombstones exceeds half the size of the keys array
(i.e. more tombstones than valid keys), the tombstones are garbage
collected.  Tombstones at the start and end of the keys array are always
removed.

To avoid re-indexing, the index uses an offset to account for elements
added or removed from the start of the keys list.

Even without much threshold tuning, this *significantly* speeds up
deletion for large hashes, while incurring minimal overhead for hashes
that never delete (and thus never add indexing overhead).